fix: correctly export next pages router files so they don't break client bundle
