+++ packages/frames.js/package.json
@@ -209,12 +209,22 @@
     },
     "./next/pages-router": {
       "import": {
-        "types": "./dist/next/pages-router.d.ts",
-        "default": "./dist/next/pages-router.js"
+        "types": "./dist/next/pages-router/index.d.ts",
+        "default": "./dist/next/pages-router/index.js"
       },
       "require": {
-        "types": "./dist/next/pages-router.d.cts",
-        "default": "./dist/next/pages-router.cjs"
+        "types": "./dist/next/pages-router/index.d.cts",
+        "default": "./dist/next/pages-router/index.cjs"
+      }
+    },
+    "./next/pages-router/client": {
+      "import": {
+        "types": "./dist/next/pages-router/client.d.ts",
+        "default": "./dist/next/pages-router/client.js"
+      },
+      "require": {
+        "types": "./dist/next/pages-router/client.d.cts",
+        "default": "./dist/next/pages-router/client.cjs"
       }
     },
     "./express": {
